requirements.txt
Django==1.10.5
django-filter==1.0.1
django-tenant-schemas==1.6.11
djangorestframework==3.5.3
Markdown==2.6.8
psycopg2==2.6.2
